+++ DelayAndSum.c
@@ -10,7 +10,12 @@ double timeDelay(float Tx, float Ty, float Tz,
                 float Rx, float Ry, float Rz, 
                 double IDx, double IDy, double IDz)
 {
-    return (sqrt(pow((Rx - IDx), 2) + pow((Ry - IDy), 2) + pow((Rz - IDz), 2)) + sqrt(pow((Tx - IDx), 2) + pow((Ty - IDy), 2) + pow((Tz - IDz), 2))) / (C / sqrt(RELATIVE_PERMITTIVITY));
+    #ifdef USE_PROPORTIONAL_PERMITTIVITY
+    return (sqrt(pow((Rx - IDx), 2) + pow((Ry - IDy), 2) + pow((Rz - IDz), 2)) + sqrt(pow((Tx - IDx), 2) + pow((Ty - IDy), 2) + pow((Tz - IDz), 2))) / (C / (BOUNDARY_RADIUS * sqrt(INNER_PERMITTIVITY)/(ANTENNA_RADIUS * sqrt(OUTER_PERMITTIVITY))));
+    #else
+    return (sqrt(pow((Rx - IDx), 2) + pow((Ry - IDy), 2) + pow((Rz - IDz), 2)) + sqrt(pow((Tx - IDx), 2) + pow((Ty - IDy), 2) + pow((Tz - IDz), 2))) / (C / (sqrt(RELATIVE_PERMITTIVITY)));
+    #endif
+
 }
 
 void delayAndSum(int chan[NUMBER_OF_CHANNELS][2],
@@ -27,11 +32,11 @@ void delayAndSum(int chan[NUMBER_OF_CHANNELS][2],
             for (int f = 0; f < FREQ_MAX; f++) // Loop over the number of frequensies 
             {
                 // Set image point coodinates and data
+                long double complex IQData = iq[f][m]; // IQ data for point over the channel and frequnecy
                 double IDx = imagingDomain[r][0];      // x
                 double IDy = imagingDomain[r][1];      // y
                 double IDz = Z_HEIGHT;                 // z
-                long double complex IQData = iq[f][m]; // IQ data for point over the channel and frequnecy
-
+                
                 // Receiving antenna coodinates
                 int RAnt = chan[m][0];      // Antenna ID
                 float Rx = antLoc[RAnt][0]; // x